added makefile, moved code in src folder, added build folder to git ignore, fixed namespace errors, renamed Chess_game to not include a space
