--- Chess Game.cpp
@@ -1,33 +1,0 @@
-// Chess Game.cpp : This file contains the 'main' function. Program execution begins and ends there.
-// Authors: Luke/WinterEkiska
-
-#include <iostream>
-#include <string>
-#include "Piece_Classes.h"
-#
-
-/*int main() {
-    std::cout << "Hello World!\n";
-}*/
-
-// Run program: Ctrl + F5 or Debug > Start Without Debugging menu
-// Debug program: F5 or Debug > Start Debugging menu
-
-// Tips for Getting Started: 
-//   1. Use the Solution Explorer window to add/manage files
-//   2. Use the Team Explorer window to connect to source control
-//   3. Use the Output window to see build output and other messages
-//   4. Use the Error List window to view errors
-//   5. Go to Project > Add New Item to create new code files, or Project > Add Existing Item to add existing code files to the project
-//   6. In the future, to open this project again, go to File > Open > Project and select the .sln file
-
-int main() {
-    const int BOARD_SIZE = 8;
-    Piece_Classes::king king;  //starts on e1, [5,1]
-    
-    //king.printLocation();
-    //std::cout << "printing king location: " << (int)king.getLocation();
-    
-
-    return 0;
-}
