--- board.cpp
@@ -1,20 +1,0 @@
-#pragma once
-
-#include <iostream>
-#include <string>
-#include "Piece_Classes.h"
-#include "board.h"
-
-
-const int SIZE = 8;
-
-int Board::boardInit(){
-	
-
-	return 2;
-}
-
-int Board::boardPrint() {
-
-	return 3;
-}
