+++ src/board.cpp
@@ -1,0 +1,17 @@
+
+#include "board.h"
+
+#include <iostream>
+#include <string>
+
+#include "Piece_Classes.h"
+
+namespace Board {
+
+const int SIZE = 8;
+
+int board::boardInit() { return 2; }
+
+int board::boardPrint() { return 3; }
+
+};  // namespace Board
